Trying to update for Dart version r14406

+++ example/click/click_demo.dart
@@ -62,7 +62,7 @@ class ClickDemo{
     window.requestAnimationFrame(_onFrame);
   }
 
-  void _onFrame(int highResTime){
+  void _onFrame(double highResTime){
     _stage.draw();
     requestFrame();
   }
